% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PCA_score.R
\name{calculate_CPA_score}
\alias{calculate_CPA_score}
\title{Calculate PCA Score from Segmentation}
\usage{
calculate_CPA_score(sample_segmentation)
}
\arguments{
\item{sample_segmentation}{Segmentation Dataframe}
}
\value{
Numeric; CPA score
}
\description{
Calculate PCA Score from Segmentation
}
