% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{create_blacklist_gr}
\alias{create_blacklist_gr}
\title{Create Blacklist regions GRanges object}
\usage{
create_blacklist_gr(blacklist_files)
}
\arguments{
\item{blacklist_files}{Character; Filepath to file containing blacklist regions}
}
\value{
GRanges object of blacklist regions
}
\description{
Create Blacklist regions GRanges object
}
