% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PoN.R
\name{create_PoN}
\alias{create_PoN}
\title{Create Panel-of-Normal (PoN) Rdata}
\usage{
create_PoN(list_rdsfiles, file, overwrite = TRUE)
}
\arguments{
\item{list_rdsfiles}{Character; a file contains paths to Profile.Rdata per line}

\item{file}{Character; Output file (.rds)}

\item{overwrite}{Logical: Overwrite existing if result already existed}
}
\value{
Null
}
\description{
Create Panel-of-Normal (PoN) Rdata
}
