% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{fragment_dist}
\alias{fragment_dist}
\title{Get insert-size distribution table}
\usage{
fragment_dist(readbam_bin, maximum_length = 600, minimum_length = 20)
}
\arguments{
\item{readbam_bin}{SampleBam Object from function read_bamfile}

\item{maximum_length}{Int; Maximum length of fragment. cfDNA fragment longer than this value will not be considered; Default 600}

\item{minimum_length}{Int; Minimum length of fragment. cfDNA fragment shorter than this value will not be considered;  Default 20}
}
\value{
Distribution table of fragment length
}
\description{
Get insert-size distribution table
}
