% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{get_fragment_profile}
\alias{get_fragment_profile}
\title{Getting fragment-length information
Function extract insert-size (fragment-length)}
\usage{
get_fragment_profile(
  readbam_bin,
  short_range = c(100, 150),
  long_range = c(151, 250),
  maximum_length = 600,
  minimum_length = 20
)
}
\arguments{
\item{readbam_bin}{list}

\item{short_range}{numeric vector}

\item{long_range}{numeric vector}

\item{maximum_length}{numeric}

\item{minimum_length}{numeric}
}
\value{
list
}
\description{
Getting fragment-length information
Function extract insert-size (fragment-length)
}
\examples{

}
