% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cnv.R
\name{get_segment_bysolution}
\alias{get_segment_bysolution}
\title{Return CNV segmentation result from given all CNV solutions}
\usage{
get_segment_bysolution(solution, sample_segmentation, SL_distance_df)
}
\arguments{
\item{solution}{solution dataframe}

\item{sample_segmentation}{Segmeantion dataframe}

\item{SL_distance_df}{Distance matrix}
}
\value{
list of segmentation per solution
}
\description{
Return CNV segmentation result from given all CNV solutions
}
