% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_solution_table.R
\name{get_solution_table}
\alias{get_solution_table}
\title{Get summarised table of cnv solutions}
\usage{
get_solution_table(cnv_solutions)
}
\arguments{
\item{cnv_solutions}{cnvcalling result from function call_cnv.R}
}
\value{
Dataframe of solution table
}
\description{
Get summarised table of cnv solutions
}
