% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zscore_profile.R
\name{get_zscore_profile}
\alias{get_zscore_profile}
\title{Transform SLRatio with PoN Fragment profile}
\usage{
get_zscore_profile(fragment_profile, pon_profile_rds)
}
\arguments{
\item{fragment_profile}{Sample Profile}

\item{pon_profile_rds}{Path to PoN rds file}
}
\value{
Dataframe of robust transformed SLratio
}
\description{
Transform SLRatio with PoN Fragment profile
}
