% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{GRCh2UCSCGRanges}
\alias{GRCh2UCSCGRanges}
\title{Convert GRCh chromosome format to UCSC style}
\usage{
GRCh2UCSCGRanges(which)
}
\arguments{
\item{which}{GRanges;}
}
\value{
GRanges; GRanges after chromosome format conversion
}
\description{
Convert GRCh chromosome format to UCSC style
}
