% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{if_exist_baifile}
\alias{if_exist_baifile}
\title{Check if bai file exist from given bam}
\usage{
if_exist_baifile(bamfile)
}
\arguments{
\item{bamfile}{Character; Path to sample bamfile}
}
\value{
Boolean if the bai file exist
}
\description{
Check if bai file exist from given bam
}
