% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{if_ucsc_chrformat}
\alias{if_ucsc_chrformat}
\title{Check UCSC chromosomes format for input bam file}
\usage{
if_ucsc_chrformat(bamfile_path)
}
\arguments{
\item{bamfile_path}{Character; Path to sample bamfile}
}
\value{
Boolean; if the input bam file is UCSC format, chr prefix
}
\description{
Check UCSC chromosomes format for input bam file
}
