% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{make_density_table}
\alias{make_density_table}
\title{Make Fragment-length density table}
\usage{
make_density_table(readbam_bin, minimum_length, maximum_length)
}
\arguments{
\item{readbam_bin}{List; A list containing SampleBam object/objects from the read_bamfile function}

\item{minimum_length}{numeric;}

\item{maximum_length}{numeric}
}
\value{
data.frame
}
\description{
Make Fragment-length density table
}
