% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv_solution.R
\name{overlap_bin_with_segment}
\alias{overlap_bin_with_segment}
\title{Overlap and merge bin data frame with segmentation dataframe}
\usage{
overlap_bin_with_segment(per_bin_profile, sample_segmentation)
}
\arguments{
\item{per_bin_profile}{bin dataframe}

\item{sample_segmentation}{segmentation dataframe}
}
\value{
dataframe of overlapping bin and segmentation
}
\description{
Overlap and merge bin data frame with segmentation dataframe
}
