% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv_solution.R
\name{plot_cnv_solution}
\alias{plot_cnv_solution}
\title{Plot Fragment-length profile with CNV calling result}
\usage{
plot_cnv_solution(cnvcall, selected_solution = 1, ylim = c(-30, 30))
}
\arguments{
\item{cnvcall}{solution results from call_cnv function}

\item{selected_solution}{solution rank to plot}

\item{ylim}{Vector of 2 Int; ylim of plot (default c(-20,20))}
}
\value{
ggplot object plot Genomics CNV profile of selected solution
}
\description{
Plot Fragment-length profile with CNV calling result
}
