% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_matrix.R
\name{plot_distance_matrix}
\alias{plot_distance_matrix}
\title{Plot Distance Matrix from CNVCalling}
\usage{
plot_distance_matrix(cnvcall)
}
\arguments{
\item{cnvcall}{cnvcalling result from function call_cnv.R}
}
\value{

}
\description{
Plot Distance Matrix from CNVCalling
}
