% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isize_distribution.R
\name{plot_isize_distribution}
\alias{plot_isize_distribution}
\title{Plot Fragment-length Distribution}
\usage{
plot_isize_distribution(
  fragment_profile,
  sample_name = "Patient cfDNA",
  xlim = 500
)
}
\arguments{
\item{fragment_profile}{list}

\item{sample_name}{charactor}

\item{xlim}{numeric}
}
\value{
distribution plot
}
\description{
Plot Fragment-length Distribution
}
