% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sl_ratio.R
\name{plot_sl_ratio}
\alias{plot_sl_ratio}
\title{Plot Short/Long-fragment Ratio}
\usage{
plot_sl_ratio(
  fragment_profile,
  chrTotalLength_file = "hg19_chrTotalLength.tsv",
  ylim = c(0, 0.4)
)
}
\arguments{
\item{fragment_profile}{list}

\item{chrTotalLength_file}{character}

\item{ylim}{numeric vector}
}
\value{
plot
}
\description{
Plot Short/Long-fragment Ratio
}
