% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PoN.R
\name{read_PoN_files}
\alias{read_PoN_files}
\title{Read Fragment Profile from a list of rds file}
\usage{
read_PoN_files(list_rdsfiles)
}
\arguments{
\item{list_rdsfiles}{path to file containing list of rds file}
}
\value{
list containing content of rds file
}
\description{
Read Fragment Profile from a list of rds file
}
