% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_control_density.R
\name{save_as_control_density}
\alias{save_as_control_density}
\title{Save Fragment Profile for distribution comparison}
\usage{
save_as_control_density(readbam_bin, fragment_profile, file)
}
\arguments{
\item{readbam_bin}{variable from read_bamfile}

\item{fragment_profile}{Fragment profile}

\item{file}{Filename}
}
\value{

}
\description{
Save Fragment Profile for distribution comparison
}
