% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fragment_profile.R
\name{save_fragment_profile}
\alias{save_fragment_profile}
\title{Save Fragment Profile as RDS}
\usage{
save_fragment_profile(fragment_profile, output_dir, overwrite = TRUE)
}
\arguments{
\item{fragment_profile}{Fragment profile variable from function get_fragment_profile}

\item{output_dir}{Path to output directory}

\item{overwrite}{Boolean; TRUE to overwrite existing file}
}
\value{
Null
}
\description{
Save Fragment Profile as RDS
}
