% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentByPSCB.R
\name{segmentByPSCB}
\alias{segmentByPSCB}
\title{Segmentation data with PSCBS}
\usage{
segmentByPSCB(sample_transformed_sl)
}
\arguments{
\item{sample_transformed_sl}{dataframe of z-transformed SLRatio}
}
\value{
Dataframe of segmentation result
}
\description{
Segmentation data with PSCBS
}
