% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_control_density.R
\name{set_control_density}
\alias{set_control_density}
\title{Setting control distribution profile}
\usage{
set_control_density(control_rds)
}
\arguments{
\item{control_rds}{Name of control rds file in ext directory}
}
\value{
Null
}
\description{
Setting control distribution profile
}
