% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{test_isize_KolmogorovSmirnov}
\alias{test_isize_KolmogorovSmirnov}
\title{KolmogorovSmirnov test for insert size}
\usage{
test_isize_KolmogorovSmirnov(control_insert_size, sample_insert_size)
}
\arguments{
\item{control_insert_size}{Vector of insert size of a control sample}

\item{sample_insert_size}{Vector of insert size of a testing sample}
}
\value{
KS.Test result
}
\description{
KolmogorovSmirnov test for insert size
}
