% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{test_KolmogorovSmirnov}
\alias{test_KolmogorovSmirnov}
\title{Test fragment Profile with Kolmogorov-Smirnov}
\usage{
test_KolmogorovSmirnov(control_fragment_profile, sample_fragment_profile)
}
\arguments{
\item{control_fragment_profile}{Fragment profile of control sample}

\item{sample_fragment_profile}{Fragment profile of testing sample}
}
\value{
list
}
\description{
Test fragment Profile with Kolmogorov-Smirnov
}
