% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{UCSC2GRChSampleBam}
\alias{UCSC2GRChSampleBam}
\title{Convert UCSC chromosome format to GRCh style from a list of alignment information}
\usage{
UCSC2GRChSampleBam(sample.bam)
}
\arguments{
\item{sample.bam}{list of alignment information from function read_bamfile}
}
\value{
List; list of alignment information after conversion
}
\description{
Convert UCSC chromosome format to GRCh style from a list of alignment information
}
