% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util.bias_correct}
\alias{util.bias_correct}
\title{Correct GC Bias readcount}
\usage{
util.bias_correct(readcount, bias)
}
\arguments{
\item{readcount}{numeric}

\item{bias}{numeric}
}
\value{
numeric
}
\description{
Correct GC Bias readcount
}
