% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util.load_control_density_table}
\alias{util.load_control_density_table}
\title{Load control fragment profile}
\usage{
util.load_control_density_table(
  control_rds = "healthycontrol.fragmentprofile.RDS",
  label = "Healthy Control"
)
}
\arguments{
\item{control_rds}{Name of healthy fragment rds file in extdata directory}

\item{label}{Label of the sample}
}
\value{
List containing fragment length distribution table and all insert size
}
\description{
Load control fragment profile
}
