% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.file_exists}
\alias{utils.file_exists}
\title{Check if file/files given do exist}
\usage{
utils.file_exists(files_vector)
}
\arguments{
\item{files_vector}{Vector of filepath}
}
\value{
Boolean; if all files exist.
}
\description{
Check if file/files given do exist
}
