% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_solution_table.R
\name{write_solution_table}
\alias{write_solution_table}
\title{Write solution table to file}
\usage{
write_solution_table(cnvcall, outputfile)
}
\arguments{
\item{cnvcall}{cnvcalling result from function call_cnv.R}

\item{outputfile}{Character; Destination file name}
}
\description{
Write solution table to file
}
