% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zscore_profile.R
\name{zscore_transform}
\alias{zscore_transform}
\title{zscore_transform transforms SLRatio profile into z-score}
\usage{
zscore_transform(per_bin_profile)
}
\arguments{
\item{per_bin_profile}{SampleFragment from function get_fragment_profile}
}
\value{
dataframe of z-score per bin
}
\description{
zscore_transform transforms SLRatio profile into z-score
}
