% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PCA_score.R
\name{calculate_CPA_score}
\alias{calculate_CPA_score}
\title{Calculate PCA Score from Segmentation}
\usage{
calculate_CPA_score(sample_segmentation)
}
\arguments{
\item{sample_segmentation}{Segmentation Dataframe}
}
\value{
Numeric; CPA score
}
\description{
Calculate PCA Score from Segmentation
}
\examples{
### Loading example SampleBam file
example_file =  system.file("extdata","example_patientcfDNA.RDS",package = "cfdnakit")
sample_bambin <- readRDS(example_file)
### Example PoN
PoN_rdsfile = system.file("extdata","ex.PoN.rds",package = "cfdnakit")
sample_profile <- get_fragment_profile(sample_bambin,sample_id = "Patient1")

sample_zscore = get_zscore_profile(sample_profile,PoN_rdsfile)

sample_zscore_segment = segmentByPSCB(sample_zscore)

calculate_CPA_score(sample_zscore_segment)

}
