% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cnv.R
\name{call_cnv}
\alias{call_cnv}
\title{Call Copy-number Variation from SLRatio and segmentation}
\usage{
call_cnv(sample_segmentation, sample_zscore)
}
\arguments{
\item{sample_segmentation}{segmentation dataframe from segmentByPSCBS}

\item{sample_zscore}{zscore dataframe}
}
\value{

}
\description{
Call Copy-number Variation from SLRatio and segmentation
}
