% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PoN.R
\name{create_PoN}
\alias{create_PoN}
\title{Create Panel-of-Normal (PoN) Rdata}
\usage{
create_PoN(
  list_rdsfiles,
  output_dir,
  filename = "PoN_fragment_profiles",
  overwrite = TRUE
)
}
\arguments{
\item{list_rdsfiles}{a file contains paths to Profile.Rdata per line}

\item{output_dir}{Output directory}

\item{filename}{Output filename}

\item{overwrite}{Logical: Overwrite existing if result already existed}
}
\value{

}
\description{
Create Panel-of-Normal (PoN) Rdata
}
