% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PoN.R
\name{create_PoN}
\alias{create_PoN}
\title{Create Panel-of-Normal (PoN) Rdata}
\usage{
create_PoN(list_rdsfiles, file, overwrite = TRUE)
}
\arguments{
\item{list_rdsfiles}{Character; a file contains paths to Profile.Rdata per line}

\item{file}{Character; Output file (.rds)}

\item{overwrite}{Logical: Overwrite existing if result already existed}
}
\value{
Null
}
\description{
Create Panel-of-Normal (PoN) Rdata
}
\examples{
healthy.1 = system.file("extdata","ex.healthy1.rds",package = "cfdnakit")
healthy.2 = system.file("extdata","ex.healthy2.rds",package = "cfdnakit")

list_rdsfile = "reference_healthy.listfile"
fileConn<-file(list_rdsfile)
writeLines(c(healthy.1,healthy.2), fileConn)
close(fileConn)
create_PoN(list_rdsfile,"temp.ex.PoN.rds")

file.remove("temp.ex.PoN.rds")
}
