% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{filter_read_on_blacklist}
\alias{filter_read_on_blacklist}
\title{Filter out reads on blacklist regions}
\usage{
filter_read_on_blacklist(sample_bin)
}
\arguments{
\item{sample_bin}{Variable sample bin from read_bamfile function}
}
\value{
Sample read per bin after filtering out read on balck list regions
}
\description{
Filter out reads on blacklist regions
}
