% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_profile.R
\name{make_density_table}
\alias{make_density_table}
\title{Make Fragment-length density table}
\usage{
make_density_table(isize_vector, minimum_length, maximum_length)
}
\arguments{
\item{isize_vector}{numeric}

\item{minimum_length}{numeric}

\item{maximum_length}{numeric}
}
\value{
data.frame
}
\description{
Make Fragment-length density table
}
