% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isize_distribution.R
\name{plot_isize_distribution}
\alias{plot_isize_distribution}
\title{Plot Fragment-length Distribution}
\usage{
plot_isize_distribution(fragment_profile, xlim = 500)
}
\arguments{
\item{fragment_profile}{list}

\item{xlim}{numeric}
}
\value{

}
\description{
Plot Fragment-length Distribution
}
