% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sl_ratio.R
\name{plot_sl_ratio}
\alias{plot_sl_ratio}
\title{Plot Short/Long-fragment Ratio}
\usage{
plot_sl_ratio(fragment_profile, ylim = c(0, 0.4), genome = "hg19")
}
\arguments{
\item{fragment_profile}{list}

\item{ylim}{plot y-axis limit}

\item{genome}{Character; version of reference genome (default hg19)}
}
\value{
plot
}
\description{
Plot Short/Long-fragment Ratio
}
