% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transformed_sl.R
\name{plot_transformed_sl}
\alias{plot_transformed_sl}
\title{Plot z-tranformed Short/Long-fragment Ratio}
\usage{
plot_transformed_sl(
  sample_transformed_sl,
  sample_segment_df = NULL,
  ylim = c(-30, 30)
)
}
\arguments{
\item{sample_transformed_sl}{Dataframe z-transformed SLRatio from get_zscore_profile}

\item{sample_segment_df}{Dataframe segmenation from segmentByPSCB}

\item{ylim}{plot y-axis limit}
}
\value{
Genome-wide plot of z-transformed SLRatio
}
\description{
Plot z-tranformed Short/Long-fragment Ratio
}
