% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{read_bamfile}
\alias{read_bamfile}
\title{Read a bam file
Read a bam file from give path. Alignment and sequencing read information
will be binned into non-overlapping size}
\usage{
read_bamfile(bamfile_path, binsize = 1000, target_bedfile = NULL)
}
\arguments{
\item{bamfile_path}{Path to sample bamfile}

\item{binsize}{Size of non-overlapping windows in KB. Max = 1000; Default 1000}

\item{target_bedfile}{Path to exon/target bedfile; Default NULL}
}
\value{
List of fragment information per bin
}
\description{
Read a bam file
Read a bam file from give path. Alignment and sequencing read information
will be binned into non-overlapping size
}
