% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bamfile.R
\name{read_bamfile}
\alias{read_bamfile}
\title{Read a bam file
Read a bam file from give path. Alignment and sequencing read information
will be binned into non-overlapping size}
\usage{
read_bamfile(bamfile_path, binsize)
}
\arguments{
\item{bamfile_path}{character}

\item{binsize}{numeric}
}
\value{

}
\description{
Read a bam file
Read a bam file from give path. Alignment and sequencing read information
will be binned into non-overlapping size
}
