% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.file_exists}
\alias{utils.file_exists}
\title{Check if file/files given do exist}
\usage{
utils.file_exists(files_vector)
}
\arguments{
\item{files_vector}{filepath/s}
}
\value{
Boolean; if all files exist.
}
\description{
Check if file/files given do exist
}
\examples{
example_file =  system.file("extdata","example_patientcfDNA.RDS",package = "cfdnakit")
utils.file_exists(example_file)
}
